[PATCH] Fix older error en base.html and finish docker-compose for server and web
